V1 of DepartmentManagment api

+++ DepartmentMangment/Authentication.Presentation/Controllers/DepartmentController.cs
@@ -7,9 +7,11 @@ using DepartManagment.Application.Models.Employee;
 using DepartManagment.Application.Queries.Departments.GetDepartment;
 using DepartManagment.Application.Queries.Departments.GetDepartments;
 using DepartManagment.Domain.Entities.ApplicationUser;
+using DepartManagment.Domain.Entities.ApplicationUser.Enums;
 using ErrorOr;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System;
@@ -32,11 +34,12 @@ public class DepartmentController : BaseController
 {
     private readonly ISender _sender;
     private readonly ILogger<DepartmentController> _logger;
-
-    public DepartmentController(ISender sender, ILogger<DepartmentController> logger)
+    private readonly UserManager<Employee> _userManager;
+    public DepartmentController(ISender sender, ILogger<DepartmentController> logger, UserManager<Employee> userManager)
     {
         _sender = sender;
         _logger = logger;
+        _userManager = userManager;
     }
 
 
@@ -44,7 +47,15 @@ public class DepartmentController : BaseController
     /*    [HasPermission(Permission.CanDeployDepartments)]*/
     public async Task<IActionResult> CreareDepartment(DepartmentCreateUpdateModel departmentCreateUpdateModel)
     {
-      /*  string username = User.FindFirst(c => c.Type == ClaimTypes.NameIdentifier)!.Value;*/
+        string Admin = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+        Employee AdminEmployee = await _userManager.FindByNameAsync(Admin);
+        if (AdminEmployee.Role != Role.Admin)
+        {
+            var error = new Results();
+            error.AddErrorMessages("You are not authorized");
+            return Unauthorized(error);
+        }
+
 
         var DeployDepartmentCommand = new DeployDepartmentCommand(departmentCreateUpdateModel.Name,departmentCreateUpdateModel.ManagerId);
         ErrorOr<DepartmentViewModel> results = await _sender.Send(DeployDepartmentCommand);
@@ -58,6 +69,15 @@ public class DepartmentController : BaseController
 /*    [Authorize(AuthenticationSchemes = "Bearer")]*/
     public async Task<IActionResult> AddEmployeeToDepartment(Guid departmentId, string employeeId)
     {
+        string Admin = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+        Employee AdminEmployee = await _userManager.FindByNameAsync(Admin);
+        if (AdminEmployee.Role != Role.Admin)
+        {
+            var error = new Results();
+            error.AddErrorMessages("You are not authorized");
+            return Unauthorized(error);
+        }
+
         var command = new AddEmployeeToDepartmentCommand(departmentId, employeeId);
         ErrorOr<EmployeeViewModel> result = await _sender.Send(command);
         return result.Match(
@@ -70,6 +90,15 @@ public class DepartmentController : BaseController
     [HttpPut("{id}")]
     public async Task<IActionResult> UpdateDepartment(Guid id, [FromBody] UpdateDepartmentModel departmentUpdateModel)
     {
+        string Admin = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+        Employee AdminEmployee = await _userManager.FindByNameAsync(Admin);
+        if (AdminEmployee.Role != Role.Admin)
+        {
+            var error = new Results();
+            error.AddErrorMessages("You are not authorized");
+            return Unauthorized(error);
+        }
+
         if (!ModelState.IsValid)
         {
             return BadRequest(ModelState);
@@ -87,6 +116,15 @@ public class DepartmentController : BaseController
     [HttpGet("{id}")]
     public async Task<IActionResult> GetDepartmentById(Guid id)
     {
+        string Admin = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+        Employee AdminEmployee = await _userManager.FindByNameAsync(Admin);
+        if (AdminEmployee.Role != Role.Admin)
+        {
+            var error = new Results();
+            error.AddErrorMessages("You are not authorized");
+            return Unauthorized(error);
+        }
+
         var query = new GetDepartmentByIdQuery(id);
         ErrorOr<DepartmentViewModel> result = await _sender.Send(query);
 
@@ -103,6 +141,15 @@ public class DepartmentController : BaseController
     [HttpGet]
     public async Task<IActionResult> GetDepartments(int pageNumber, int pageSize)
     {
+        string Admin = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+        Employee AdminEmployee = await _userManager.FindByNameAsync(Admin);
+        if (AdminEmployee.Role != Role.Admin)
+        {
+            var error = new Results();
+            error.AddErrorMessages("You are not authorized");
+            return Unauthorized(error);
+        }
+
         var query = new GetAllDepartmentsQuery(pageNumber, pageSize);
         ErrorOr<List<DepartmentsViewModel>> result = await _sender.Send(query);
 
@@ -118,6 +165,15 @@ public class DepartmentController : BaseController
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteDepartment(Guid id)
     {
+        string Admin = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+        Employee AdminEmployee = await _userManager.FindByNameAsync(Admin);
+        if (AdminEmployee.Role != Role.Admin)
+        {
+            var error = new Results();
+            error.AddErrorMessages("You are not authorized");
+            return Unauthorized(error);
+        }
+
         var command = new DeleteDepartmentCommand(id);
         var results = await _sender.Send(command);
